# api/public.openapi.yaml
openapi: 3.0.3
info:
    title: WEBYCP Public API
    version: 0.7.0
    description: Public REST API used by the WEBYCP web application.
paths:
...
    /api/v1/accounts:
        get:
            operationId: listAccounts
            summary: List accessible hosting accounts
            security:
                - cookieAuth: []
            parameters:
                - $ref: "#/components/parameters/Page"
                - $ref: "#/components/parameters/PageSize"
            responses:
                "200":
                    description: Hosting accounts.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/AccountListResponse"
                "401":
                    $ref: "#/components/responses/UnauthorizedError"
                "400":
                    $ref: "#/components/responses/BadRequestError"
                "500":
                    $ref: "#/components/responses/InternalError"
        post:
            operationId: createAccount
            summary: Create and queue provisioning for a hosting account
            security:
                - cookieAuth: []
            parameters:
                - $ref: "#/components/parameters/CSRFToken"
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/CreateAccountRequest"
            responses:
                "202":
                    description: Account provisioning queued.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/AccountJobResponse"
                "400":
                    $ref: "#/components/responses/BadRequestError"
                "401":
                    $ref: "#/components/responses/UnauthorizedError"
                "403":
                    $ref: "#/components/responses/ForbiddenError"
                "404":
                    $ref: "#/components/responses/NotFoundError"
                "409":
                    $ref: "#/components/responses/ConflictError"
                "422":
                    $ref: "#/components/responses/ValidationError"
                "500":
                    $ref: "#/components/responses/InternalError"
    /api/v1/accounts/{accountId}:
        patch:
            operationId: setAccount
            summary: Enable or disable a hosting account
            security: [{ cookieAuth: [] }]
            parameters:
                - $ref: "#/components/parameters/AccountID"
                - $ref: "#/components/parameters/CSRFToken"
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/UpdateEnabledRequest"
            responses:
                "202":
                    description: Account operation queued.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/AccountJobResponse"
                "400": { $ref: "#/components/responses/BadRequestError" }
                "401": { $ref: "#/components/responses/UnauthorizedError" }
                "403": { $ref: "#/components/responses/ForbiddenError" }
                "404": { $ref: "#/components/responses/NotFoundError" }
                "409": { $ref: "#/components/responses/ConflictError" }
                "500": { $ref: "#/components/responses/InternalError" }
        delete:
            operationId: deleteAccount
            summary: Delete an empty hosting account and quarantine its home
            security: [{ cookieAuth: [] }]
            parameters:
                - $ref: "#/components/parameters/AccountID"
                - $ref: "#/components/parameters/CSRFToken"
            responses:
                "202":
                    description: Account deletion queued.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/AccountJobResponse"
                "401": { $ref: "#/components/responses/UnauthorizedError" }
                "403": { $ref: "#/components/responses/ForbiddenError" }
                "404": { $ref: "#/components/responses/NotFoundError" }
                "409": { $ref: "#/components/responses/ConflictError" }
                "500": { $ref: "#/components/responses/InternalError" }
...
    /api/v1/nodes:
        get:
            operationId: listNodes
            summary: List managed nodes
            security:
                - cookieAuth: []
            responses:
                "200":
                    description: Managed nodes.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/NodeListResponse"
                "401":
                    $ref: "#/components/responses/UnauthorizedError"
                "500":
                    $ref: "#/components/responses/InternalError"
    /api/v1/nodes/{nodeId}/probe:
        post:
            operationId: probeNode
            summary: Queue a node connectivity probe
            security:
                - cookieAuth: []
            parameters:
                - $ref: "#/components/parameters/NodeID"
                - $ref: "#/components/parameters/CSRFToken"
            responses:
                "202":
                    description: Probe queued.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Job"
                "401":
                    $ref: "#/components/responses/UnauthorizedError"
                "403":
                    $ref: "#/components/responses/ForbiddenError"
                "404":
                    $ref: "#/components/responses/NotFoundError"
                "500":
                    $ref: "#/components/responses/InternalError"
    /api/v1/jobs:
        get:
            operationId: listJobs
            summary: List recent jobs
            security:
                - cookieAuth: []
            parameters:
                - $ref: "#/components/parameters/Page"
                - $ref: "#/components/parameters/PageSize"
            responses:
                "200":
                    description: Recent jobs.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/JobListResponse"
                "400":
                    $ref: "#/components/responses/BadRequestError"
                "401":
                    $ref: "#/components/responses/UnauthorizedError"
                "500":
                    $ref: "#/components/responses/InternalError"
...
components:
    securitySchemes:
        cookieAuth:
            type: apiKey
            in: cookie
            name: webycp_session
    parameters:
        Page:
            name: page
            in: query
            schema:
                type: integer
                minimum: 1
                default: 1
        PageSize:
            name: size
            in: query
            schema:
                type: integer
                minimum: 1
                maximum: 100
                default: 10
        CSRFToken:
            name: X-CSRF-Token
            in: header
            required: true
            schema:
                type: string
        NodeID:
            name: nodeId
            in: path
            required: true
            schema:
                type: string
...
        AccountID:
            name: accountId
            in: path
            required: true
            schema: { type: string }
...
    responses:
        BadRequestError:
            description: The request is malformed.
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ErrorResponse"
        UnauthorizedError:
            description: Authentication is required.
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ErrorResponse"
        ForbiddenError:
            description: The request is not allowed.
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ErrorResponse"
        NotFoundError:
            description: The resource does not exist.
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ErrorResponse"
        ConflictError:
            description: The request conflicts with current state.
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ErrorResponse"
        ValidationError:
            description: The request is invalid.
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ErrorResponse"
        InternalError:
            description: An internal error occurred.
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ErrorResponse"
    schemas:
...
        Node:
            type: object
            additionalProperties: false
            required: [id, name, kind, endpoint, status, createdAt, updatedAt]
            properties:
                id:
                    type: string
                name:
                    type: string
                kind:
                    type: string
                    enum: [local, remote]
                endpoint:
                    type: string
                status:
                    type: string
                    enum: [unknown, online, offline]
                lastSeenAt:
                    type: string
                    format: date-time
                    nullable: true
                createdAt:
                    type: string
                    format: date-time
                updatedAt:
                    type: string
                    format: date-time
        Account:
            type: object
            additionalProperties: false
            required: [id, nodeId, name, systemUser, status, enabled, createdAt, updatedAt]
            properties:
                id:
                    type: string
                nodeId:
                    type: string
                name:
                    type: string
                systemUser:
                    type: string
                status:
                    type: string
                    enum: [pending, active, disabled, error]
                enabled:
                    type: boolean
                createdAt:
                    type: string
                    format: date-time
                updatedAt:
                    type: string
                    format: date-time
        CreateAccountRequest:
            type: object
            additionalProperties: false
            required: [name, nodeId, packageId]
            properties:
                name:
                    type: string
                    minLength: 2
                    maxLength: 80
                nodeId:
                    type: string
                packageId:
                    type: string
...
        PackageLimits:
            type: object
            additionalProperties: false
            required:
                [websites, domains, aliases, databases, databaseUsers, scheduledTasks, backupPlans, backupRetention]
            properties:
                websites: { type: integer, format: int64, minimum: 0, maximum: 1000000 }
                domains: { type: integer, format: int64, minimum: 0, maximum: 1000000 }
                aliases: { type: integer, format: int64, minimum: 0, maximum: 1000000 }
                databases: { type: integer, format: int64, minimum: 0, maximum: 1000000 }
                databaseUsers: { type: integer, format: int64, minimum: 0, maximum: 1000000 }
                scheduledTasks: { type: integer, format: int64, minimum: 0, maximum: 1000000 }
                backupPlans: { type: integer, format: int64, minimum: 0, maximum: 1000000 }
                backupRetention: { type: integer, format: int64, minimum: 1, maximum: 100 }
        PackageUsage:
            type: object
            additionalProperties: false
            required: [websites, domains, aliases, databases, databaseUsers, scheduledTasks, backupPlans]
            properties:
                websites: { type: integer, format: int64, minimum: 0 }
                domains: { type: integer, format: int64, minimum: 0 }
                aliases: { type: integer, format: int64, minimum: 0 }
                databases: { type: integer, format: int64, minimum: 0 }
                databaseUsers: { type: integer, format: int64, minimum: 0 }
                scheduledTasks: { type: integer, format: int64, minimum: 0 }
                backupPlans: { type: integer, format: int64, minimum: 0 }
        Package:
            type: object
            additionalProperties: false
            required: [id, name, limits, accountCount, createdAt, updatedAt]
            properties:
                id: { type: string }
                name: { type: string }
                limits: { $ref: "#/components/schemas/PackageLimits" }
                accountCount: { type: integer, format: int64, minimum: 0 }
                createdAt: { type: string, format: date-time }
                updatedAt: { type: string, format: date-time }
...
        AccountOverview:
            type: object
            additionalProperties: false
            required: [id, nodeId, name, systemUser, status, enabled, createdAt, updatedAt, package, usage]
            properties:
                id: { type: string }
                nodeId: { type: string }
                name: { type: string }
                systemUser: { type: string }
                status:
                    type: string
                    enum: [pending, active, disabled, error]
                enabled: { type: boolean }
                createdAt: { type: string, format: date-time }
                updatedAt: { type: string, format: date-time }
                package: { $ref: "#/components/schemas/Package" }
                usage: { $ref: "#/components/schemas/PackageUsage" }
...
        AccountListResponse:
            type: object
            additionalProperties: false
            required: [items, pagination]
            properties:
                items:
                    type: array
                    items:
                        $ref: "#/components/schemas/AccountOverview"
                pagination:
                    $ref: "#/components/schemas/Pagination"
        Pagination:
            type: object
            additionalProperties: false
            required: [page, size, totalItems, totalPages]
            properties:
                page:
                    type: integer
                    minimum: 1
                size:
                    type: integer
                    minimum: 1
                    maximum: 100
                totalItems:
                    type: integer
                    format: int64
                    minimum: 0
                totalPages:
                    type: integer
                    minimum: 0
        AccountJobResponse:
            type: object
            additionalProperties: false
            required: [account, job]
            properties:
                account:
                    $ref: "#/components/schemas/Account"
                job:
                    $ref: "#/components/schemas/Job"
        UpdateEnabledRequest:
            type: object
            additionalProperties: false
            required: [enabled]
            properties:
                enabled: { type: boolean }
...
        NodeListResponse:
            type: object
            additionalProperties: false
            required: [items]
            properties:
                items:
                    type: array
                    items:
                        $ref: "#/components/schemas/Node"
        Job:
            type: object
            additionalProperties: false
            required: [id, kind, status, payload, attempts, maxAttempts, createdAt]
            properties:
                id:
                    type: string
                nodeId:
                    type: string
                    nullable: true
                userId:
                    type: string
                    nullable: true
                kind:
                    type: string
                status:
                    type: string
                    enum: [queued, running, succeeded, failed]
                payload:
                    type: object
                    additionalProperties: true
                attempts:
                    type: integer
                    format: int64
                maxAttempts:
                    type: integer
                    format: int64
                error:
                    type: string
                    nullable: true
                createdAt:
                    type: string
                    format: date-time
                startedAt:
                    type: string
                    format: date-time
                    nullable: true
                finishedAt:
                    type: string
                    format: date-time
                    nullable: true
...
        JobListResponse:
            type: object
            additionalProperties: false
            required: [items, pagination]
            properties:
                items:
                    type: array
                    items:
                        $ref: "#/components/schemas/Job"
                pagination:
                    $ref: "#/components/schemas/Pagination"
...
        ErrorResponse:
            type: object
            additionalProperties: false
            required: [code, message]
            properties:
                code:
                    type: string
                message:
                    type: string
                fields:
                    type: object
                    additionalProperties:
                        type: string